refactor: sonic
